+++ lib/eve_online.rb
@@ -156,6 +156,7 @@ require 'eve_online/esi/models/group'
 require 'eve_online/esi/models/graphic'
 require 'eve_online/esi/models/system'
 require 'eve_online/esi/models/planet'
+require 'eve_online/esi/models/planets'
 
 require 'eve_online/esi/models/killmail_short'