+++ lib/eve_online/esi/models/planets.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module EveOnline
+  module ESI
+    module Models
+      class Planets < Base
+        def planets
+          output = []
+
+          options.each do |planet|
+            output << Planet.new(planet)
+          end
+
+          output
+        end
+      end
+    end
+  end
+end